Conflicts:
	DataManager.xcodeproj/project.xcworkspace/xcuserdata/mrj.xcuserdatad/UserInterfaceState.xcuserstate

+++ DataManager/DataManager/HttpManager.h
@@ -56,7 +56,7 @@
 - (AFHTTPRequestOperation *)postRequestToUrl:(NSString *)url
                                       params:(NSDictionary *)params
                                        files:(NSArray *)files
-                                     process:(void (^)(double writedBytes, double totalBytes))process
+                                     process:(void (^)(int64_t writedBytes, int64_t totalBytes))process
                                     complete:(void (^)(BOOL successed, NSDictionary *result))complete;
 /*
  filePath : 下载文件的存储路径
@@ -66,7 +66,7 @@
 - (AFHTTPRequestOperation *)downloadFileWithUrl:(NSString *)url
                                          params:(NSDictionary *)params
                                        filePath:(NSString *)filePath
-                                        process:(void (^)(double readBytes, double totalBytes))process
+                                        process:(void (^)(int64_t readBytes, int64_t totalBytes))process
                                        complete:(void (^)(BOOL successed, NSDictionary *response))complete;
 
 @end
